+++ src/main/java/com/mz/sistema/gestao/escolar/servico/TransferenciaServico.java
@@ -18,7 +18,7 @@ public interface TransferenciaServico {
 	public List<Transferencia> receberTransferenciasPorIdEscolaPorAno(String pesquisa, Long idEscola, Integer ano);
 	public List<Transferencia> obterTransferenciasPorIdEscolaPorAno(String pesquisa, Long idEscola, Integer ano);
 	public List<Transferencia> receberTransferenciasPorIdEscolaPorAnoComEstadoFalse(Long idEscola, Integer ano);
-	public Transferencia obterTransferenciasExistente(Long idEscolaOrigem, Long idEscolaDestino, Integer idMatricula);
+	public Transferencia obterTransferenciasExistente(Long idEscolaOrigem, Long idEscolaDestino, Long idMatricula);
 	public Long obterAlunosdosQueTransferiramNestaEscola(Long idEscolaOrigem, Integer ano, boolean finalizada);
 	public Long obterAlunosdosQueForamTransferidosParaEstaEscola(Long idEscolaDestino, Integer ano, boolean finalizada);
 }
